Expose createDocument and delegateToSchema
Currently part of unofficial, unsupported and undocumented API.
